# Commit of improved IndependentSolver::getInitalValues().