--- a/include/klee/Constraints.h
+++ b/include/klee/Constraints.h
@@ -65,6 +65,11 @@ public:
     return constraints == other.constraints;
   }
   
+  ref<Expr> get(unsigned index) const{
+    return constraints[index];
+  }
+
+
 private:
   std::vector< ref<Expr> > constraints;