--- a/lib/Core/Executor.cpp
+++ b/lib/Core/Executor.cpp
@@ -268,8 +268,8 @@ namespace {
 
   cl::opt<unsigned>
   MaxForksTerminate("max-forks-terminate",
-		  cl::desc("Only fork this many times and then dump states (default=-1 (off))"),
-          cl::init(~0u));
+  		cl::desc("Only fork this many times and then dump states (default=-1 (off))"),
+  		cl::init(~0u));
 }
 
 
@@ -2933,7 +2933,7 @@ ref<Expr> Executor::replaceReadWithSymbolic(ExecutionState &state,
   // and return it.
   
   static unsigned id;
-  const Array *array = new Array("rrws_arr" + llvm::utostr(++id), 
+  const Array *array = Array::CreateArray("rrws_arr" + llvm::utostr(++id),
                                  Expr::getMinBytesForWidth(e->getWidth()));
   ref<Expr> res = Expr::createTempRead(array, e->getWidth());
   ref<Expr> eq = NotOptimizedExpr::create(EqExpr::create(e, res));
@@ -3272,7 +3272,7 @@ void Executor::executeMakeSymbolic(ExecutionState &state,
     while (!state.arrayNames.insert(uniqueName).second) {
       uniqueName = name + "_" + llvm::utostr(++id);
     }
-    const Array *array = new Array(uniqueName, mo->size);
+    const Array *array = Array::CreateArray(uniqueName, mo->size);
     bindObjectInState(state, mo, false, array);
     state.addSymbolic(mo, array);