… lib/Solver/IndependenceAnalysis.cpp
… lib/Solver/IndependentSolver.cpp